move unit tests from src to tests dirs
