+++ program-libs/interface/Cargo.toml
@@ -12,3 +12,6 @@ solana = ["dep:solana-instruction", "dep:solana-pubkey"]
 borsh = { workspace = true }
 solana-instruction = { workspace = true, optional = true }
 solana-pubkey = { workspace = true, optional = true, features = ["curve25519"] }
+
+[dev-dependencies]
+solana-pubkey = { workspace = true }
